Modules branch with minor changes to make it stable
This commit virtually merges Modules branch (under development) with master branch, after editting a few lines to make it work as if Modules wasn't there. The intention is to avoid conflicts when merging master into Modules in the future; that is, to keep both branches as similar as possible as long as old (tested) classes and methods are still used in master branch.